Persist the stream before getting finish ACK
For some cases, we don't ever persist whatever we get, and
thus makes it hard to debug what the build was doing before
it hung completely.

This should persist the existing output every X minutes (set
to 5 minutes) so that we can get a better understanding of the
failure conditions for certain cases.